% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSubgroupQuality.R
\name{testSubgroupSetsQuality}
\alias{testSubgroupSetsQuality}
\title{Function compares different sets of patterns
  and verify how many of them a related to the goal class}
\usage{
testSubgroupSetsQuality(sets = list(), labels = c(),
  qfunc = function(labels, qfunc.opt) {     NULL }, qfunc.opt = NULL,
  mode = "both", n.bstrp = 100)
}
\arguments{
\item{sets}{a list/vector of different sets, every set is a list of different pattern indices}

\item{labels}{a list/vector of target attribute(s) information. The attributes are aggregate by the function qfunc}

\item{qfunc}{is a function that takes labels of the target group a measure it relevancy to this attribute}

\item{qfunc.opt}{is an arbitrary data needed by qfunc}

\item{mode}{is the mode for computing extream values it can be one of 'small','large','both'.}

\item{n.bstrp}{is the number bootstrap trials}
}
\value{
For every set returns the probability that the quality value is within the confidence interval
}
\description{
Function compares different sets of patterns
  and verify how many of them a related to the goal class
}
\examples{

sets = list(list(SetName="TMP", Extents=list(c(1,5,7),c(2,3,4,5))))
labels = rbinom(10,1,0.5)
qfunc=function(labels, qfunc.opt=NULL) {
  return(mean(labels))
}
n.bstrp=100

testSubgroupSetsQuality(sets,labels,qfunc,NULL,n.bstrp)
}
